requirements.txt
matplotlib
openpyxl
meteostat
pandas
pdfminer.six
tabula-py
